requirements.txt
autopep8==2.0.0
click==8.1.3
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.9.1
python-dotenv==0.21.0
tomli==2.0.1
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.10.0
